% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm.R
\name{bmrarm}
\alias{bmrarm}
\title{MCMC sampler to implement a bmrarm model}
\usage{
bmrarm(
  formula,
  data,
  ordinal_outcome,
  time_var,
  patient_var,
  random_slope = T,
  ar_cov = TRUE,
  nsim = 1000,
  burn_in = 100,
  thin = 10,
  seed = 14,
  sig_prior = 1e+05,
  sd_vec = c(0.15, 0.3, rep(0.2, 4)),
  N_burn_trunc = 10,
  prior_siw_uni = c(0.2, 5),
  prior_siw_df = NULL,
  prior_siw_scale_mat = NULL
)
}
\arguments{
\item{formula}{an object of class "formula"; a symbolic description of the model to be fitted}

\item{data}{a dataframe containing outcome variables, covariates, and a patient or subject identifier}

\item{ordinal_outcome}{a character string, which contains the variable name for the ordinal outcome}

\item{time_var}{a character string, which contains the variable name for time indexing;
this must be integer valued and should indicate the observation number.}

\item{patient_var}{a character string, which contains the variable name for patient indexing}

\item{ar_cov}{logical, indicating use of an autoregressive error term. The default is TRUE}

\item{nsim}{positive integer, number of iterations with default of 1000}

\item{burn_in}{positive integer, number of iterations to remove with default of 100}

\item{thin}{positive integer, specifies the period of saving samples. Default of 10}

\item{seed}{positive integer, seed for random number generation}

\item{sig_prior}{scalar, variance term for prior on the beta coefficients}

\item{sd_vec}{numeric vector, containing standard deviations for Metroplis Hastings proposals, either of length 4 or 6.
The values are associated with the MH-within-gibbs step, the ar term, and the random effects}

\item{N_burn_trunc}{integer, number of burn-in draws from the truncated multivariate normal Gibbs sampler}

\item{prior_siw_uni}{prior bounds for the uniform distribution associated with the SIW prior}

\item{prior_siw_df}{degrees of freedom for the SIW prior.
The default is 1 + number of random effects for a single person (2 or 4).}

\item{prior_siw_scale_mat}{scale matrix for the SIW prior. The default is an identity matrix}

\item{random_Slope}{logical, indicating use of random slopes. The default is TRUE}
}
\value{
bmrarm
}
\description{
MCMC sampler to implement a bmrarm model
}
