% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrarm.R
\name{bmrarm_fc_missing}
\alias{bmrarm_fc_missing}
\title{Function to sample the latent continuous values}
\usage{
bmrarm_fc_missing(y, z, X, Z_kron, cur_draws, samp_info)
}
\arguments{
\item{y}{matrix of latent and observed continuous observations}

\item{z}{vector of the ordinal outcomes}

\item{X}{design matrix}

\item{Z_kron}{design matrix for random effects}

\item{cur_draws}{list of current parameter values}

\item{samp_info}{list of internal info used for sampling}
}
\value{
list
}
\description{
Function to sample the latent continuous values
}
