% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrarm.R
\name{bmrarm_fc_patient_siw}
\alias{bmrarm_fc_patient_siw}
\title{Function to sample the random effects and the associated covariance matrix}
\usage{
bmrarm_fc_patient_siw(
  y,
  z,
  X,
  cur_draws,
  samp_info,
  Z_kron,
  prior_siw_uni,
  prior_siw_df,
  prior_siw_scale_mat
)
}
\arguments{
\item{y}{matrix of latent and observed continuous observations}

\item{z}{vector of the ordinal outcomes}

\item{X}{design matrix}

\item{cur_draws}{list of current parameter values}

\item{samp_info}{list of internal info used for sampling}

\item{Z_kron}{design matrix for random effects}

\item{prior_siw_uni}{prior bounds for the uniform distribution associated with the SIW prior}

\item{prior_siw_df}{degrees of freedom for the SIW prior}

\item{prior_siw_scale_mat}{scale matrix for the SIW prior}
}
\value{
list
}
\description{
Function to sample the random effects and the associated covariance matrix
}
