% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrarm.R
\name{bmrarm_mh_ar}
\alias{bmrarm_mh_ar}
\title{Full conditional draws of the regression coefficients}
\usage{
bmrarm_mh_ar(y, X, Z_kron, cur_draws, samp_info)
}
\arguments{
\item{subject_effects}{matrix of patient specific intercepts}

\item{sigma}{residual covariance matrix}

\item{prior_alpha}{prior term for scale}
}
\value{
scalar
}
\description{
Full conditional draws of the regression coefficients
}
