% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_simulations.R
\docType{data}
\name{bmrarm_simulations}
\alias{bmrarm_simulations}
\title{Generated datasets for bmrarm simulations}
\format{
Two lists of data frames. One list contains data generated with an autoregressive term, the other without:
\describe{
\item{y1}{true continuous values for the ordinal outcome, unused for modeled}
\item{y_ord}{ordinal outcome, obtained from discretizing y1}
\item{y2}{observed continuous outcome}
\item{pat_idx}{patient identifier, integer}
\item{time}{time since baseline, integer}
\item{data_type}{identifies if data was used for training or testing. Only applicable to ar_data_list data frames as all data was training data for the slope_data_list data frames.}
}
}
\usage{
bmrarm_simulations
}
\description{
A list of datasets used in the simulated bmrarm examples.
}
\keyword{datasets}
