% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{bmrarm_start}
\alias{bmrarm_start}
\title{Create lists, vectors, and matrices for storage and passing to functions}
\usage{
bmrarm_start(env)
}
\arguments{
\item{env}{parent environment}
}
\description{
Create lists, vectors, and matrices for storage and passing to functions
}
