% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx.R
\name{bmrvarx}
\alias{bmrvarx}
\title{PX-DA MCMC routine to implement a bmrvarx model}
\usage{
bmrvarx(
  formula,
  data,
  ordinal_outcomes,
  sig_prior = 1e+06,
  nsim = 1000,
  burn_in = 100,
  thin = 10,
  seed = 14,
  max_iter_rej = 500,
  N_burn_trunc = 10
)
}
\arguments{
\item{formula}{an object of class "formula"; a symbolic description of the model to be fitted}

\item{data}{a dataframe containing outcome variables, covariates, and a patient or subject identifier}

\item{ordinal_outcomes}{a character string containing the names of the ordinal outcomes}

\item{sig_prior}{scalar, prior variance on the regression coefficients}

\item{nsim}{positive integer, number of iterations with default of 1000}

\item{burn_in}{positive integer, number of iterations to remove with default of 100}

\item{thin}{positive integer, specifiers the period of saving samples. Default of 10}

\item{seed}{positive integer, seed for random number generation}

\item{max_iter_rej}{maximum number of rejection algorithm attempts for multivariate truncated normal}

\item{N_burn_trunc}{integer, number of burn-in draws from the truncated multivariate normal Gibbs sampler}
}
\value{
bmrvarx
}
\description{
PX-DA MCMC routine to implement a bmrvarx model
}
