% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_helper_functions.R
\name{cond_mean_part}
\alias{cond_mean_part}
\title{Extract conditional mean corresponsing to ordinal components}
\usage{
cond_mean_part(y_current, mean_vec, pre_calc_mat, ord_loc)
}
\arguments{
\item{y_current}{current continous outcome values}

\item{mean_vec}{vector of mean values}

\item{pre_calc_mat}{a pre-calculated matrix}

\item{ord_loc}{number of ordinal outcomes}
}
\value{
scalar
}
\description{
Extract conditional mean corresponsing to ordinal components
}
