% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds.R
\name{cont_to_cat}
\alias{cont_to_cat}
\title{Discretize a latent value using cutpoints}
\usage{
cont_to_cat(y, cut_points)
}
\arguments{
\item{y}{scalar continuous value}

\item{cut_points}{vector of cutpoints}
}
\description{
Discretize a latent value using cutpoints
}
