% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_helper_functions.R
\name{create_storage}
\alias{create_storage}
\title{Create storage for all samples in the time series setting}
\usage{
create_storage(env)
}
\arguments{
\item{env}{parent enviroment}
}
\description{
Create storage for all samples in the time series setting
}
