% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrarm.R
\name{dmatrix_normal_log}
\alias{dmatrix_normal_log}
\title{Function to calculate values for bmrarm_mh_ar accept reject step}
\usage{
dmatrix_normal_log(resid_mat, cur_draws, samp_info, sig_list)
}
\arguments{
\item{resid_mat}{matrix of residuals}

\item{cur_draws}{list of current parameter values}

\item{samp_info}{list of covariance matrices and determinants}
}
\value{
scalar
}
\description{
Function to calculate values for bmrarm_mh_ar accept reject step
}
