% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrvarx.R
\name{fc_cuts}
\alias{fc_cuts}
\title{Full conditional draws for the threshold parameters}
\usage{
fc_cuts(y, z, N_cat, upper_cut_limit)
}
\arguments{
\item{y}{matrix of multivariate observations}

\item{z}{matrix of ordinal outcomes}

\item{N_cat}{vector of positive integers, number of categories for each ordinal outcome}
}
\value{
list
}
\description{
Full conditional draws for the threshold parameters
}
