% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_da.R
\name{fc_cuts_da}
\alias{fc_cuts_da}
\title{Full conditional draws for the threshold parameters}
\usage{
fc_cuts_da(y, z, N_cat)
}
\arguments{
\item{y}{matrix of multivariate observations}

\item{z}{matrix of ordinal outcomes}

\item{N_cat}{vector of positive integers, number of categories for each ordinal outcome}
}
\value{
list
}
\description{
Full conditional draws for the threshold parameters
}
