% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrvarx.R
\name{fc_sigma_theta_tilde}
\alias{fc_sigma_theta_tilde}
\title{Function to draw regression coefficients and covariance matrix}
\usage{
fc_sigma_theta_tilde(y, X, prior_precision, y_orig)
}
\arguments{
\item{y}{matrix of multivariate continuous observations}

\item{X}{design matrix}

\item{prior_precision}{prior precision matrix}

\item{y_orig}{unscaled version of y}
}
\value{
list
}
\description{
Function to draw regression coefficients and covariance matrix
}
