% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{first_cut_prob}
\alias{first_cut_prob}
\title{Probability calculation as part of the MH-within-Gibbs step (Cowles' algorithm)}
\usage{
first_cut_prob(samp_info, cuts, cuts_tmp)
}
\arguments{
\item{samp_info}{vector of proposed current cutpoints}

\item{cuts}{vector of current cutpoints}
}
\description{
Probability calculation as part of the MH-within-Gibbs step (Cowles' algorithm)
}
