% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_ar_errors}
\alias{gen_ar_errors}
\title{Generate data for longitudinal simulation}
\usage{
gen_ar_errors(
  N = 7,
  N_pat = 48,
  seed = 10,
  unequal = FALSE,
  slope = F,
  ar_cov = T
)
}
\arguments{
\item{N}{Mean time points per patient, drawn from a poisson}

\item{N_pat}{number of patients}

\item{seed}{seed to set, default of 10}
}
\description{
Generate data for longitudinal simulation
}
