% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_bmrvarx}
\alias{gen_bmrvarx}
\title{Generate data for time series simulation}
\usage{
gen_bmrvarx(N = 600, seed = 10, N_param = 3)
}
\arguments{
\item{N}{Number of timepoints}

\item{seed}{seed to set, default of 10}
}
\description{
Generate data for time series simulation
}
