% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{get_dev}
\alias{get_dev}
\title{Get deviance for a single set of parameter values, used to calculate DIC}
\usage{
get_dev(
  y,
  X,
  z,
  Z_kron,
  pat_idx,
  pat_idx_long,
  beta,
  sig_list,
  cuts,
  pat_eff,
  miss_mat,
  samp_info,
  marginal
)
}
\arguments{
\item{y}{matrix of latent and observed continuous observations}

\item{X}{design matrix}

\item{z}{vector of the ordinal outcomes}

\item{Z_kron}{design matrix for random effects}

\item{pat_idx}{vector to index patients for both outcomes}

\item{beta}{matrix of beta values}

\item{sig_list}{list of covariance matrix values}

\item{cuts}{vector of cutpoints}

\item{pat_eff}{matrix of patient effects}

\item{miss_mat}{matrix to identify locations of missing values}

\item{samp_info}{list of interal info used for sampling}

\item{marginal}{logical, indicates of mDIC of cDIC is to be calculated}
}
\value{
scalar
}
\description{
Get deviance for a single set of parameter values, used to calculate DIC
}
