% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{get_dev_ar3}
\alias{get_dev_ar3}
\title{Get DIC}
\usage{
get_dev_ar3(
  y,
  X,
  z,
  Z_kron,
  pat_idx,
  pat_idx_long,
  beta,
  sig_list,
  cuts,
  pat_eff,
  miss_mat,
  samp_info
)
}
\arguments{
\item{y_current}{current continous outcome values}

\item{mean_vec}{vector of mean values}

\item{pre_calc_mat}{a pre-calculated matrix}

\item{ord_loc}{number of ordinal outcomes}
}
\value{
scalar
}
\description{
Get DIC
}
