% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{get_DIC}
\alias{get_DIC}
\title{Get marginal or conditional DIC for a bmrarm object}
\usage{
get_DIC(samps, marginal = TRUE)
}
\arguments{
\item{samps}{a bmrarm object}

\item{marginal}{logical, used marginal (the default) or conditional DIC}
}
\value{
vector
}
\description{
Get marginal or conditional DIC for a bmrarm object
}
