% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{get_loocv}
\alias{get_loocv}
\title{Get DIC}
\usage{
get_loocv(
  samps,
  cv_locs,
  z_true,
  y_true,
  full_X,
  full_Z_kron,
  data,
  ar = FALSE
)
}
\arguments{
\item{y_current}{current continous outcome values}

\item{mean_vec}{vector of mean values}

\item{pre_calc_mat}{a pre-calculated matrix}

\item{ord_loc}{number of ordinal outcomes}
}
\value{
scalar
}
\description{
Get DIC
}
