% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{get_lppd_ar}
\alias{get_lppd_ar}
\title{Get DIC}
\usage{
get_lppd_ar(
  y,
  X,
  z,
  Z_kron,
  pat_idx,
  pat_idx_long,
  beta,
  sigma,
  cuts,
  pat_eff,
  miss_mat,
  samp_info,
  cv_locs,
  ar_val,
  data
)
}
\arguments{
\item{y_current}{current continous outcome values}

\item{mean_vec}{vector of mean values}

\item{pre_calc_mat}{a pre-calculated matrix}

\item{ord_loc}{number of ordinal outcomes}
}
\value{
scalar
}
\description{
Get DIC
}
