% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds.R
\name{get_pred}
\alias{get_pred}
\title{Generate posterior predictive draws for bmrarm object. Not intended for public use.}
\usage{
get_pred(samps, new_X = NULL)
}
\arguments{
\item{samps}{bmrarm object with additional structure for new time points}

\item{new_X}{new design matrix}
}
\value{
list
}
\description{
Generate posterior predictive draws for bmrarm object. Not intended for public use.
}
