% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds.R
\name{get_preds_bmrvarx}
\alias{get_preds_bmrvarx}
\title{Full conditional draws of the latent continuous values}
\usage{
get_preds_bmrvarx(
  bmrvarx_obj,
  steps_ahead = 5,
  X,
  seed = 15,
  get_ord_preds = T
)
}
\arguments{
\item{bmrvarx_obj}{an object returned from bmrvarx()}

\item{steps_ahead}{number of observations to forcast, done recursively}

\item{X}{matrix of covariates to make forecasts with, number of rows must
equal stes_ahead}
}
\value{
matrix
}
\description{
Full conditional draws of the latent continuous values
}
