% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_helper_functions.R
\name{get_sampling_info}
\alias{get_sampling_info}
\title{Starting values for cut points}
\usage{
get_sampling_info(env)
}
\arguments{
\item{N_cats}{vector, number of categories for each ordinal variable}
}
\description{
Starting values for cut points
}
