% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{get_sig_list}
\alias{get_sig_list}
\title{Create list of covariance matrices}
\usage{
get_sig_list(cur_draws, samp_info)
}
\arguments{
\item{cur_draws}{cur_draws list of current parameter values}

\item{samp_info}{list of internal info used for sampling}
}
\description{
Create list of covariance matrices
}
