% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{get_sig_list_marg}
\alias{get_sig_list_marg}
\title{Create list of covariance matrices, only used in posterior summaries}
\usage{
get_sig_list_marg(cur_draws, samp_info)
}
\arguments{
\item{cur_draws}{cur_draws list of current parameter values}

\item{samp_info}{list of internal info used for sampling}
}
\description{
Create list of covariance matrices, only used in posterior summaries
}
