% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{get_sym_mat}
\alias{get_sym_mat}
\title{Create AR matrix}
\usage{
get_sym_mat(cur_draws, dist_mat)
}
\arguments{
\item{cur_draws}{list of the current parameter draws}

\item{dist_mat}{symmetric distance matrix with 0 down the diagonal}
}
\description{
Create AR matrix
}
