% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_helper_functions.R
\name{pre_calc}
\alias{pre_calc}
\title{Extract conditional mean corresponding to ordinal components}
\usage{
pre_calc(cov_mat, samp_locs)
}
\arguments{
\item{cov_mat}{covariance matrix based on first, last, or middle obs}

\item{samp_locs}{list of unique locations to be sampled}
}
\value{
scalar
}
\description{
Extract conditional mean corresponding to ordinal components
}
