% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{pre_calc_ar}
\alias{pre_calc_ar}
\title{Obtain conditional mean and covariance for multivariate normals}
\usage{
pre_calc_ar(cov_mat, locs)
}
\arguments{
\item{cov_mat}{covariance matrix based on first, last, or middle obs}

\item{locs}{vector of locations to be sampled}
}
\value{
scalar
}
\description{
Obtain conditional mean and covariance for multivariate normals
}
