% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{start_cuts}
\alias{start_cuts}
\title{Starting values for cut points}
\usage{
start_cuts(N_cats, fixed = 1)
}
\arguments{
\item{N_cats}{vector, number of categories for each ordinal variable}
}
\description{
Starting values for cut points
}
