% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summ}
\alias{summ}
\title{PX-DA MCMC routine to sample from HBMRVAR model}
\usage{
summ(x)
}
\arguments{
\item{x}{A matrix of posterior draws}
}
\description{
PX-DA MCMC routine to sample from HBMRVAR model
}
