% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bmrarm}
\alias{summary.bmrarm}
\title{Summary function for a bmrarm object}
\usage{
\method{summary}{bmrarm}(x, digits = 3)
}
\arguments{
\item{x}{A bmrarm object}

\item{digits}{scalar, number of decimial places to round to}
}
\description{
Summary function for a bmrarm object
}
