% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bmrvarx}
\alias{summary.bmrvarx}
\title{Summary function for a bmrvarx object}
\usage{
\method{summary}{bmrvarx}(x, digits = 3)
}
\arguments{
\item{x}{A bmrarm object}

\item{digits}{scalar, number of decimial places to round to}
}
\description{
Summary function for a bmrvarx object
}
