% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds.R
\name{summary_preds_bmrarm}
\alias{summary_preds_bmrarm}
\title{Generate predictions for bmrarm object. Not intended for usage.}
\usage{
summary_preds_bmrarm(preds, truth)
}
\arguments{
\item{samps}{bmrarm object with additional structure for new time points.}
}
\value{
list
}
\description{
Generate predictions for bmrarm object. Not intended for usage.
}
