% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrvarx.R
\name{tmvn_gibbs_rej}
\alias{tmvn_gibbs_rej}
\title{Gibbs step for truncated multivariate normal}
\usage{
tmvn_gibbs_rej(
  y_current,
  mean,
  lower,
  upper,
  locs,
  loc_length,
  pre_calcs,
  max_iter,
  N_burn_trunc
)
}
\arguments{
\item{y_current}{continuous outcome values}

\item{mean}{mean vector of multivariate normal}

\item{lower}{scalar, lower truncation bound}

\item{upper}{scalar, upper truncation bound}

\item{locs}{vector of locations to sample}

\item{loc_length}{length of locs}

\item{pre_calcs}{list of conditional means and covariance matrices}

\item{max_iter}{integer, maximum number of rejection sampler attempts}

\item{N_burn_trunc}{integer, number of Gibbs burn-in iterations}
}
\value{
list
}
\description{
Gibbs step for truncated multivariate normal
}
