% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrarm.R
\name{bmrarm_fc_patient_siw}
\alias{bmrarm_fc_patient_siw}
\title{Full conditional draws of the latent continuous values}
\usage{
bmrarm_fc_patient_siw(y, z, X, cur_draws, samp_info, sep_sig)
}
\arguments{
\item{y}{matrix of multivariate observations}

\item{z}{matrix of ordinal voutcomes}

\item{samp_info}{information for which locations to sample}

\item{sig}{covariance matrix for the VAR process}

\item{sig0}{for the initial values}

\item{M}{transition matrix for the VAR(1) component}

\item{cuts}{current threshold values}

\item{miss_mat}{locations of missing values}

\item{num_iter}{current iteration number}
}
\value{
matrix
}
\description{
Full conditional draws of the latent continuous values
}
