% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrarm.R
\name{bmrarm_fc_sig_beta}
\alias{bmrarm_fc_sig_beta}
\title{Function to sample regression coefficients and covariance matrix}
\usage{
bmrarm_fc_sig_beta(y, X, Z_kron, cur_draws, samp_info)
}
\arguments{
\item{y}{matrix of latent and observed continuous observations}

\item{X}{design matrix}

\item{Z_kron}{design matrix for random effects}

\item{cur_draws}{list of current parameter values}

\item{samp_info}{list of internal info used for sampling}
}
\value{
list
}
\description{
Function to sample regression coefficients and covariance matrix
}
