% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_da.R
\name{bmrvarx_da}
\alias{bmrvarx_da}
\title{PA MCMC routine to sample from BMRVAR model}
\usage{
bmrvarx_da(
  formula,
  data,
  ordinal_outcomes = c("y_ord", "y_bin"),
  sig_prior = 1e+06,
  all_draws = FALSE,
  nsim = 1000,
  burn_in = 100,
  thin = 10,
  seed = 14,
  verbose = TRUE,
  max_iter_rej = 500,
  return_y = FALSE,
  fast = F,
  old_prior_y0 = F
)
}
\arguments{
\item{nsim}{scalar, number of iterations with default of 1000}

\item{burn_in}{scalar, number of iterations to remove with default of 50}

\item{seed}{seed to default, default of 14}

\item{y}{matrix of multivariate observations}

\item{z}{vector of disease status}

\item{df}{dataframe with patient and time point indexes}
}
\value{
mcmc
}
\description{
PA MCMC routine to sample from BMRVAR model
}
