% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_helper_functions.R
\name{create_storage}
\alias{create_storage}
\title{Create lists, vectors, and matrices for storage and passing to functions}
\usage{
create_storage(env)
}
\arguments{
\item{env}{parent environment}
}
\description{
Create lists, vectors, and matrices for storage and passing to functions
}
