% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrarm.R
\name{dmatrix_normal_log}
\alias{dmatrix_normal_log}
\title{Full conditional draws of the regression coefficients}
\usage{
dmatrix_normal_log(resid_mat, cur_draws, samp_info, sig_list)
}
\arguments{
\item{subject_effects}{matrix of patient specific intercepts}

\item{sigma}{residual covariance matrix}

\item{prior_alpha}{prior term for scale}
}
\value{
scalar
}
\description{
Full conditional draws of the regression coefficients
}
