% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrvarx.R
\name{expansion_prior}
\alias{expansion_prior}
\title{Full conditional draws of the latent continuous values}
\usage{
expansion_prior(cor_mat, N_ordinal)
}
\arguments{
\item{cor_mat}{correlation matrix from previous draw}

\item{N_ordinal}{number of ordinal outcomes}
}
\value{
matrix
}
\description{
Full conditional draws of the latent continuous values
}
