% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_da.R
\name{fc_cuts_da}
\alias{fc_cuts_da}
\title{Function to draw cutpoint parameters for the DA implementation}
\usage{
fc_cuts_da(y, z, N_cat)
}
\arguments{
\item{y}{matrix of continuous observations}

\item{z}{matrix of ordinal outcomes}

\item{N_cat}{vector of positive integers, number of categories for each ordinal outcome}
}
\value{
list
}
\description{
Function to draw cutpoint parameters for the DA implementation
}
