% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrvarx.R
\name{fc_y}
\alias{fc_y}
\title{Full conditional draws of the latent continuous values}
\usage{
fc_y(y, z, mean_mat, tmp_list, miss_mat, samp_info, num_iter)
}
\arguments{
\item{y}{matrix of multivariate observations}

\item{z}{matrix of ordinal voutcomes}

\item{miss_mat}{locations of missing values}

\item{samp_info}{information for which locations to sample}

\item{num_iter}{current iteration number}

\item{sig}{covariance matrix for the VAR process}

\item{sig0}{for the initial values}

\item{M}{transition matrix for the VAR(1) component}

\item{cuts}{current threshold values}
}
\value{
matrix
}
\description{
Full conditional draws of the latent continuous values
}
