% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{first_cut_prob}
\alias{first_cut_prob}
\title{Starting values for cut points}
\usage{
first_cut_prob(samp_info, cuts, cuts_tmp)
}
\arguments{
\item{N_cats}{vector, number of categories for each ordinal variable}
}
\description{
Starting values for cut points
}
