% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds.R
\name{get_preds_bmrarm}
\alias{get_preds_bmrarm}
\title{Get DIC}
\usage{
get_preds_bmrarm(samps)
}
\arguments{
\item{y_current}{current continous outcome values}

\item{mean_vec}{vector of mean values}

\item{pre_calc_mat}{a pre-calculated matrix}

\item{ord_loc}{number of ordinal outcomes}
}
\value{
scalar
}
\description{
Get DIC
}
