% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preds.R
\name{get_preds_bmrvarx}
\alias{get_preds_bmrvarx}
\title{Generate iterative forecasts for bmrvarx object}
\usage{
get_preds_bmrvarx(
  bmrvarx_obj,
  steps_ahead = 5,
  X,
  seed = 15,
  get_ord_preds = T
)
}
\arguments{
\item{bmrvarx_obj}{a bmrvarx object}

\item{steps_ahead}{number of observations to forcast, done recursively}

\item{X}{matrix of covariates to make forecasts with, number of rows must
equal steps_ahead}

\item{get_ord_preds}{logical, discretize latent values into ordinal predictions}
}
\value{
list
}
\description{
Generate iterative forecasts for bmrvarx object
}
