% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_helper_functions.R
\name{get_sampling_info}
\alias{get_sampling_info}
\title{Create list of sampling info to pass to functions}
\usage{
get_sampling_info(env)
}
\arguments{
\item{env}{parent environment}
}
\description{
Create list of sampling info to pass to functions
}
