% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrarm_helper_functions.R
\name{get_sig_list}
\alias{get_sig_list}
\title{Create list of covariance matrices and optionally their inverse}
\usage{
get_sig_list(cur_draws, samp_info)
}
\arguments{
\item{env}{parent environment}
}
\description{
Create list of covariance matrices and optionally their inverse
}
