% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmrvarx_helper_functions.R
\name{pre_calc}
\alias{pre_calc}
\title{Extract conditional means and variances for}
\usage{
pre_calc(cov_mat, samp_locs)
}
\arguments{
\item{cov_mat}{covariance matrix}

\item{samp_locs}{list of unique locations to be sampled}
}
\value{
scalar
}
\description{
Extract conditional means and variances for
}
