% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bmrarm}
\alias{summary.bmrarm}
\title{PX-DA MCMC routine to sample from HBMRVAR model}
\usage{
\method{summary}{bmrarm}(x, digits = 3)
}
\arguments{
\item{x}{A bmrarm object}
}
\description{
PX-DA MCMC routine to sample from HBMRVAR model
}
