% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.bmrvarx}
\alias{summary.bmrvarx}
\title{PX-DA MCMC routine to sample from HBMRVAR model}
\usage{
\method{summary}{bmrvarx}(x, digits = 3)
}
\arguments{
\item{x}{A bmrvarx object}
}
\description{
PX-DA MCMC routine to sample from HBMRVAR model
}
