% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fc_bmrvarx.R
\name{tmvn_gibbs_rej}
\alias{tmvn_gibbs_rej}
\title{Gibbs step for truncated multivariate normal}
\usage{
tmvn_gibbs_rej(
  y_current,
  mean,
  lower,
  upper,
  locs,
  loc_length,
  pre_calcs,
  max_iter,
  N_ord,
  N_burn_trunc,
  num_iter,
  num_obs
)
}
\arguments{
\item{y_current}{continous outcome values}

\item{mean}{mean vector of multivariate normal}

\item{lower}{vector of lower thresholds}

\item{upper}{vector of upper thresholds}

\item{locs}{locations to sample}

\item{sigma}{covariance matrix}

\item{pre_cals}{pre calculations, including the conditional variance}
}
\value{
vector
}
\description{
Gibbs step for truncated multivariate normal
}
